+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Improve Stackdriver log format timestamp and message field formats.
 - Use crypto random to seed global random instance if possible.
 - Allow migrate subcommand to use database names that contain dashes.
+- Add senderID param to channelIdBuild function.
 
 ### Fixed
 - Fix the registered function name for 'nk.channelIdBuild' in the JavaScript runtime.